streamlit_app/requirements_streamlit.txt
streamlit
pillow
numpy
matplotlib
tensorflow
folium
streamlit-folium
